+++ R/drdid_rc1.R
@@ -87,6 +87,12 @@ drdid_rc1 <-function(y, post, D, covariates, i.weights = NULL,
   #-----------------------------------------------------------------------------
   #Compute the Pscore by MLE
   pscore.tr <- stats::glm(D ~ -1 + int.cov, family = "binomial", weights = i.weights)
+  if(pscore.tr$converged == FALSE){
+    warning(" glm algorithm did not converge")
+  }
+  if(anyNA(pscore.tr$coefficients)){
+    stop("Propensity score model coefficients have NA components. \n Multicollinearity (or lack of variation) of covariates is a likely reason.")
+  }
   ps.fit <- as.vector(pscore.tr$fitted.values)
   # Avoid divide by zero
   ps.fit <- pmin(ps.fit, 1 - 1e-16)
@@ -94,11 +100,17 @@ drdid_rc1 <-function(y, post, D, covariates, i.weights = NULL,
   reg.coeff.pre <- stats::coef(stats::lm(y ~ -1 + int.cov,
                                          subset = ((D==0) & (post==0)),
                                          weights = i.weights))
+  if(anyNA(reg.coeff.pre)){
+    stop("Outcome regression model coefficients have NA components. \n Multicollinearity (or lack of variation) of covariates is a likely reason.")
+  }
   out.y.pre <-   as.vector(tcrossprod(reg.coeff.pre, int.cov))
   #Compute the Outcome regression for the control group at the pre-treatment period, using ols.
   reg.coeff.post <- stats::coef(stats::lm(y ~ -1 + int.cov,
                                           subset = ((D==0) & (post==1)),
                                           weights = i.weights))
+  if(anyNA(reg.coeff.post)){
+    stop("Outcome regression model coefficients have NA components. \n Multicollinearity (or lack of variation) of covariates is a likely reason.")
+  }
   out.y.post <-   as.vector(tcrossprod(reg.coeff.post, int.cov))
   # Combine the ORs
   out.y <- post * out.y.post + (1 - post) * out.y.pre