+++ R/reg_did_rc.R
@@ -91,12 +91,18 @@ reg_did_rc <-function(y, post, D, covariates, i.weights = NULL,
   reg.coeff.pre <- stats::coef(stats::lm(y ~ -1 + int.cov,
                                          subset = ((D==0) & (post==0)),
                                          weights = i.weights))
+  if(anyNA(reg.coeff.pre)){
+    stop("Outcome regression model coefficients have NA components. \n Multicollinearity of covariates is probably the reason for it.")
+  }
   out.y.pre <-   as.vector(tcrossprod(reg.coeff.pre, int.cov))
   #-----------------------------------------------------------------------------
   #Compute the Outcome regression for the control group at the pre-treatment period, using ols.
   reg.coeff.post <- stats::coef(stats::lm(y ~ -1 + int.cov,
                                           subset = ((D==0) & (post==1)),
                                           weights = i.weights))
+  if(anyNA(reg.coeff.post)){
+    stop("Outcome regression model coefficients have NA components. \n Multicollinearity (or lack of variation) of covariates is probably the reason for it.")
+  }
   out.y.post <-   as.vector(tcrossprod(reg.coeff.post, int.cov))
   #-----------------------------------------------------------------------------
   #Compute the OR DID estimators