+++ NEWS.md
@@ -1,9 +1,10 @@
-# DRDID 1.0.1
+# DRDID 1.0.1 (unreleased, only on GitHub for the moment)
   * Allows for treating covariates as factor and alike when computing DiD
+  
+  * Improve error and warning handling due to collinearity and convergence issues.
   
 # DRDID 1.0.0
   * First official version of package, functions for computing a variety of difference-in-differences (DiD) estimators for the ATT. 
-  
   * Documentation is improved compared to the devel version, including examples for every function now.
   
   * Created wrapper function `drdid`, `ordid` and `ipwdid` to implement doubly-robust, outcome regression and inverse probability weighted DID estimators.